Add some transformation logic to the scaffold mediator
The logic here was arbitrarily chosen to demonstrate that a client can
make a request to the OpenHIM. The OpenHIM will then record the request
than route that request through to the Mediator. This mediator will
then make request to the DHIS2 Demo Sierra Leone setup to get a list of
all the OrganisationUnits in XML format.
The mediator will transform the response from DHIS2 from XML to JSON and
also add another field to each facility.
This transformed data is then returned to the Client through the
OpenHIM. This give the client a record of the whole transaction and the
ability to replay a transaction.

OHM-798